services/lumina-brain/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lumina-brain"
version = "1.0.0"
description = "Lumina Brain API - Vector embedding and semantic search service"
authors = [
    { name = "Felix", email = "felix@example.com" }
]
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sentence-transformers>=2.2.0",
    "qdrant-client>=1.6.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.hatch.build.targets.wheel]
packages = ["src/lumina_brain"]

[tool.hatch.version]
path = "src/lumina_brain/__init__.py"

[project.scripts]
start = "lumina_brain.main:main"
